# swagger.yaml
openapi: 3.0.0
info:
  title: Pacific Telecom
  description: Apis para Pacific Telecom
  version: 0.1
servers:
  # - url: http://localhost:3000
  - url: http://138.197.105.129:3000
    description: Pacific Telecom Resource Server

# 1) Define the security scheme type (HTTP bearer)
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
# 2) Apply the security globally to all operations
security:
  - bearerAuth: []         # use the same name as above
paths:
...
  /urbanism/{id}:
    get:
      summary: Obtener un urbanismo por id
      tags:
         - Urbanismos
      parameters:
        - name: id
          in: path
          required: true
          description: id del urbanismo
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

  
    /urbanism/{id}:
    patch: 
      summary: Actualizar urbanismo
      tags:
        - Urbanismos
      parameters:
        - name: id
          in: path
          required: true
          description: id del urbanismo
          schema:
            type : integer
            format: int64
            minimum: 1
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"parish_id": 1,"city_id":1 ,"name": "test","households_qty": 23123, "latitude":"asdasdad", "longitude": "sadasd"}'
      responses: 
        '200':
          description: OK

      /urbanism/{id}:
    delete:
      summary: Borra un urbanismo por id
      tags:
        - Urbanismos
      parameters:
        - name: id
          in: path
          required: true
          description: id del urbanismo
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK
  /urbanism/register:
    post: 
      summary: Registrar urbanismo
      tags:
        - Urbanismos
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"parish_id": 1,"city_id":1 ,"name": "test","households_qty": 23123, "latitude":"asdasdad", "longitude": "sadasd"}'
      responses: 
        '201':
          description: OK

############################################### [REPORTE DE PAGOS] ######################################

  # /payment_report/:
  #   get:
  #     summary: Obtener todos los reportes de pago
  #     tags:
  #       - Reporte de pagos
  #     responses: 
  #       '200':
  #         description: OK
...